+++ code/Simple_Log/SimpleLog/Repository/ILogContextDelete.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SimpleLog.Repository
+{
+    internal interface ILogContextDelete
+    {
+        void Delete(Message message);
+        IEnumerable<Message> Read(Search search);
+    }
+}