+++ code/Simple_Log/TestLog/Program.cs
@@ -26,7 +26,13 @@ namespace TestLog
             string methodName = System.Reflection.MethodInfo.GetCurrentMethod().Name;
             Console.WriteLine("Start executing {0} ...", methodName);
 
-            SimpleLog.Manager logManager = new SimpleLog.Manager(SimpleLog.Constants.LOG_CONSOLE);
+            dynamic settings = new System.Dynamic.ExpandoObject();
+
+            settings.CanAddError = true;
+            settings.CanAddWarning = true;
+            settings.CanAddInfo = true;
+
+            SimpleLog.Manager logManager = new SimpleLog.Manager(SimpleLog.Constants.LOG_CONSOLE, settings);
 
             SimpleLog.Message message = new SimpleLog.Message() {
                 CreatedOn = DateTime.UtcNow,
@@ -34,6 +40,7 @@ namespace TestLog
                 Group = "Group A",
                 IdentifierName = "User_ID",
                 IdentifierValue = "User-1",
+                MessageType = "error",
                 Operation = "Edit User",
                 Owner = "AppUser"
             };
@@ -53,7 +60,13 @@ namespace TestLog
             string methodName = System.Reflection.MethodInfo.GetCurrentMethod().Name;
             Console.WriteLine("Start executing {0} ...", methodName);
 
-            SimpleLog.Manager logManager = new SimpleLog.Manager(SimpleLog.Constants.LOG_DB);
+            dynamic settings = new System.Dynamic.ExpandoObject();
+
+            settings.CanAddError = true;
+            settings.CanAddWarning = true;
+            settings.CanAddInfo = true;
+
+            SimpleLog.Manager logManager = new SimpleLog.Manager(SimpleLog.Constants.LOG_DB, settings);
 
             SimpleLog.Message message = new SimpleLog.Message()
             {
@@ -62,6 +75,7 @@ namespace TestLog
                 Group = "Group A",
                 IdentifierName = "User_ID",
                 IdentifierValue = "User-1",
+                MessageType = "error",
                 Operation = "Edit User",
                 Owner = "AppUser"
             };
@@ -81,7 +95,15 @@ namespace TestLog
             string methodName = System.Reflection.MethodInfo.GetCurrentMethod().Name;
             Console.WriteLine("Start executing {0} ...", methodName);
 
-            SimpleLog.Manager logManager = new SimpleLog.Manager(SimpleLog.Constants.LOG_TEXT_FILE, 0, SimpleLog.Helper.GenerateFileName("log"));
+            dynamic settings = new System.Dynamic.ExpandoObject();
+            
+            settings.CanAddError = true;
+            settings.CanAddWarning = true;
+            settings.CanAddInfo = true;
+
+            settings.FileName = SimpleLog.Helper.GenerateFileName("log");
+
+            SimpleLog.Manager logManager = new SimpleLog.Manager(SimpleLog.Constants.LOG_TEXT_FILE, settings);
 
             for (int i = 1; i <= 10; i++)
             {
@@ -93,6 +115,7 @@ namespace TestLog
                     Group = "Group A",
                     IdentifierName = "User_ID",
                     IdentifierValue = "User-" + i.ToString(),
+                    MessageType = "error",
                     Operation = "Edit User",
                     Owner = "AppUser"
                 };