+++ versions/release-notes.md
@@ -2,3 +2,11 @@
 ---
 - [2016-04-20]
 - First released.
+
+2.0
+---
+- [2016-04-23]
+- Allow multiple contexts. So can add message to both console and text file and DB.
+- Enhance code by removing specific properties for each repository and add dynamic settings object that carry all fields either specific or generic.
+- Allow add and delete to work on all contexts that support additions or deletions.
+- Read message in certain context.